requirements.txt
anyascii==0.3.2
apturl==0.5.2
attrs==25.1.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2025.1.31
chardet==4.0.0
charset-normalizer==3.4.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.2.2
fasteners==0.14.1
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
future==0.18.2
h11==0.14.0
h5py==3.13.0
httplib2==0.20.2
huggingface-hub==0.29.1
idna==3.10
importlib-metadata==4.6.4
imutils==0.5.4
jeepney==0.7.1
Jinja2==3.1.5
keyring==23.5.0
kiwisolver==1.4.8
langdetect==1.0.9
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
matplotlib==3.10.0
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
mpmath==1.3.0
netifaces==0.11.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
onnx==1.17.0
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
paramiko==2.9.3
pexpect==4.8.0
pillow==11.1.0
protobuf==5.29.3
ptyprocess==0.7.0
PyAutoGUI==0.9.54
pycairo==1.20.1
pyclipper==1.3.0.post6
pycups==2.0.1
PyGetWindow==0.0.9
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyMuPDF==1.25.3
PyNaCl==1.5.0
pyparsing==3.2.1
pypdfium2==4.30.1
pyperclip==1.9.0
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==1.0.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-doctr==0.11.0
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.2
RapidFuzz==3.12.1
reportlab==3.6.8
requests==2.32.3
scipy==1.15.2
SecretStorage==3.3.1
selenium==4.29.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.13.1
systemd-python==234
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
triton==3.2.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
usb-creator==0.3.7
wadllib==1.3.6
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
